pyproject.toml
[project]
name = "serenuplift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "openai>=2.13.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "tensorboardx>=2.6.4",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "xgboost>=3.1.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]

[tool.uv.pip]
extra-index-url = [
    "https://download.pytorch.org/whl/cu128"
]
